backend/pyproject.toml
[project]
name = "hrmslite-backend"
version = "0.1.0"
description = "HRMS Lite FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "cryptography>=43.0.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.uv]
dev-dependencies = []
